+++ inlang/packages/paraglide/paraglide-js/src/compiler/output-structure/message-modules.ts
@@ -15,6 +15,7 @@ export function generateMessageModules(
 		strategy: NonNullable<CompilerOptions["strategy"]>;
 		cookieName: NonNullable<CompilerOptions["cookieName"]>;
 		isServer: NonNullable<CompilerOptions["isServer"]>;
+		localStorageKey: NonNullable<CompilerOptions["localStorageKey"]>;
 		experimentalMiddlewareLocaleSplitting: NonNullable<
 			CompilerOptions["experimentalMiddlewareLocaleSplitting"]
 		>;