+++ inlang/packages/paraglide/paraglide-js/src/compiler/runtime/extract-locale-from-request.js
@@ -50,6 +50,8 @@ export const extractLocaleFromRequest = (request) => {
 			locale = _locale;
 		} else if (strat === "baseLocale") {
 			return baseLocale;
+		} else if (strat === "localStorage") {
+			continue;
 		} else {
 			throw new Error(`Unsupported strategy: ${strat}`);
 		}